Crate matrix_sdk_sled

Structs

SledCryptoStore

A sled based cryptostore.

SledStateStore

SledStateStoreBuilder

Builder for SledStateStoreBuilderConfig.

Enums

MigrationConflictStrategy

Sometimes Migrations can’t proceed without having to drop existing data. This allows you to configure, how these cases should be handled.

OpenStoreError

All the errors that can occur when opening a sled store.

Functions

make_store_config

Create a StoreConfig with an opened SledStateStore that uses the given path and passphrase.