Struct matrix_sdk_sled::SledStateStore

pub struct SledStateStore { /* private fields */ }

Implementations

impl SledStateStore

pub fn drop_v1_tables(self) -> StoreResult<()>

impl SledStateStore

pub fn builder() -> SledStateStoreBuilder

Create a SledStateStoreBuilder with default parameters.

pub async fn open_crypto_store(&self) -> Result<SledCryptoStore, OpenStoreError>

Open a SledCryptoStore that uses the same database as this store.

The given passphrase will be used to encrypt private data.

pub async fn save_changes( &self, changes: &StateChanges ) -> Result<(), SledStoreError>

pub async fn get_presence_event( &self, user_id: &UserId ) -> Result<Option<Raw<PresenceEvent>>, SledStoreError>

pub async fn get_state_event( &self, room_id: &RoomId, event_type: StateEventType, state_key: &str ) -> Result<Option<Raw<AnySyncStateEvent>>, SledStoreError>

pub async fn get_state_events( &self, room_id: &RoomId, event_type: StateEventType ) -> Result<Vec<Raw<AnySyncStateEvent>>, SledStoreError>

pub async fn get_profile( &self, room_id: &RoomId, user_id: &UserId ) -> Result<Option<MinimalStateEvent<RoomMemberEventContent>>, SledStoreError>

pub async fn get_member_event( &self, room_id: &RoomId, state_key: &UserId ) -> Result<Option<RawMemberEvent>, SledStoreError>

pub async fn get_user_ids_stream( &self, room_id: &RoomId ) -> StoreResult<impl Stream<Item = StoreResult<OwnedUserId>>>

pub async fn get_stripped_user_ids_stream( &self, room_id: &RoomId ) -> StoreResult<impl Stream<Item = StoreResult<OwnedUserId>>>

pub async fn get_invited_user_ids( &self, room_id: &RoomId ) -> StoreResult<impl Stream<Item = StoreResult<OwnedUserId>>>

pub async fn get_joined_user_ids( &self, room_id: &RoomId ) -> StoreResult<impl Stream<Item = StoreResult<OwnedUserId>>>

pub async fn get_stripped_invited_user_ids( &self, room_id: &RoomId ) -> StoreResult<impl Stream<Item = StoreResult<OwnedUserId>>>

pub async fn get_stripped_joined_user_ids( &self, room_id: &RoomId ) -> StoreResult<impl Stream<Item = StoreResult<OwnedUserId>>>

pub async fn get_room_infos( &self ) -> Result<impl Stream<Item = Result<RoomInfo, SledStoreError>>, SledStoreError>

pub async fn get_stripped_room_infos( &self ) -> Result<impl Stream<Item = Result<RoomInfo, SledStoreError>>, SledStoreError>

pub async fn get_users_with_display_name( &self, room_id: &RoomId, display_name: &str ) -> Result<BTreeSet<OwnedUserId>, SledStoreError>

pub async fn get_account_data_event( &self, event_type: GlobalAccountDataEventType ) -> Result<Option<Raw<AnyGlobalAccountDataEvent>>, SledStoreError>

pub async fn get_room_account_data_event( &self, room_id: &RoomId, event_type: RoomAccountDataEventType ) -> Result<Option<Raw<AnyRoomAccountDataEvent>>, SledStoreError>